+++ app/models/taxon_name/matrix_hooks.rb
@@ -8,14 +8,13 @@ module TaxonName::MatrixHooks
   end
 
   def in_scope_observation_matrix_row_items
-    return [] unless parent_id_changed? && parent_id_change.first
+    return [] unless parent_id_changed?
     TaxonName.find(parent_id_change.last).coordinate_observation_matrix_row_items
   end
 
   def out_of_scope_observation_matrix_row_items
-    return [] unless parent_id_changed? && parent_id_change.last
-    TaxonName.find(parent_id_change.compact.first).coordinate_observation_matrix_row_items   
+    return [] unless parent_id_changed? && parent_id_change.first
+    TaxonName.find(parent_id_change.first).coordinate_observation_matrix_row_items   
   end
 
 end
-